Moved host builder code into its own interface (IHostBuilderFactory)
