+++ src/AppMotor.CliApp/CommandLine/Hosting/MethodHostBuilderFactory.cs
@@ -1,0 +1,45 @@
+﻿#region License
+// Copyright 2021 AppMotor Framework (https://github.com/skrysmanski/AppMotor)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#endregion
+
+using System;
+
+using Microsoft.Extensions.Hosting;
+
+namespace AppMotor.CliApp.CommandLine.Hosting
+{
+    /// <summary>
+    /// A <see cref="IHostBuilderFactory"/> wrapper for a factory method/function.
+    /// </summary>
+    public class MethodHostBuilderFactory : IHostBuilderFactory
+    {
+        private readonly Func<IHostBuilder> _factoryFunc;
+
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="factoryFunc">The factory method/function</param>
+        public MethodHostBuilderFactory(Func<IHostBuilder> factoryFunc)
+        {
+            this._factoryFunc = factoryFunc;
+        }
+
+        /// <inheritdoc />
+        public IHostBuilder CreateHostBuilder()
+        {
+            return this._factoryFunc();
+        }
+    }
+}
